chore: MODULE.bazel spring cleaning + shared trace helpers
- Unpin p4_constraints: git_override → BCR 20260311.0.
- Remove gutil: pure transitive dep of p4_constraints, Bzlmod resolves it.
- Remove gmp: the bazel/gmp/ wrapper was unused; behavioral_model handles
  GMP headers internally via its own genrules.
- Remove boost pins: 4 root-level boost deps were unnecessary transitive
  pins from behavioral_model.
- Extract shared trace helpers: buildDropTrace, buildOutputTrace,
  packetIngressEvent, stageEvent deduplicated from V1Model and PSA into
  TraceHelpers.kt.
- Remove resolved and invalid REFACTORING.md entries.
