--- bazel/gmp/MODULE.bazel
@@ -1,7 +1,0 @@
-module(
-    name = "gmp",
-    version = "6.3.0",
-)
-
-bazel_dep(name = "platforms", version = "0.0.11")
-bazel_dep(name = "rules_cc", version = "0.2.17")
