+++ simulator/V1ModelArchitecture.kt
@@ -6,15 +6,12 @@ import fourward.ir.v1.StageKind
 import fourward.ir.v1.StructDecl
 import fourward.sim.v1.SimulatorProto.CloneEvent
 import fourward.sim.v1.SimulatorProto.CloneSessionLookupEvent
-import fourward.sim.v1.SimulatorProto.Drop
 import fourward.sim.v1.SimulatorProto.DropReason
 import fourward.sim.v1.SimulatorProto.Fork
 import fourward.sim.v1.SimulatorProto.ForkBranch
 import fourward.sim.v1.SimulatorProto.ForkReason
 import fourward.sim.v1.SimulatorProto.LogMessageEvent
 import fourward.sim.v1.SimulatorProto.MarkToDropEvent
-import fourward.sim.v1.SimulatorProto.PacketIngressEvent
-import fourward.sim.v1.SimulatorProto.PacketOutcome
 import fourward.sim.v1.SimulatorProto.PipelineStageEvent
 import fourward.sim.v1.SimulatorProto.TraceEvent
 import fourward.sim.v1.SimulatorProto.TraceTree
@@ -649,38 +646,8 @@ class V1ModelArchitecture : Architecture {
       )
       .build()
 
-  private fun buildDropTrace(events: List<TraceEvent>, reason: DropReason): TraceTree {
-    val outcome = PacketOutcome.newBuilder().setDrop(Drop.newBuilder().setReason(reason)).build()
-    return TraceTree.newBuilder().addAllEvents(events).setPacketOutcome(outcome).build()
-  }
-
-  private fun buildOutputTrace(events: List<TraceEvent>, port: Int, payload: ByteArray): TraceTree {
-    val output =
-      fourward.sim.v1.SimulatorProto.OutputPacket.newBuilder()
-        .setEgressPort(port)
-        .setPayload(com.google.protobuf.ByteString.copyFrom(payload))
-        .build()
-    val outcome = PacketOutcome.newBuilder().setOutput(output).build()
-    return TraceTree.newBuilder().addAllEvents(events).setPacketOutcome(outcome).build()
-  }
-
-  private fun packetIngressEvent(ingressPort: UInt): TraceEvent =
-    TraceEvent.newBuilder()
-      .setPacketIngress(PacketIngressEvent.newBuilder().setIngressPort(ingressPort.toInt()))
-      .build()
-
-  private fun stageEvent(
-    stage: PipelineStage,
-    direction: PipelineStageEvent.Direction,
-  ): TraceEvent =
-    TraceEvent.newBuilder()
-      .setPipelineStage(
-        PipelineStageEvent.newBuilder()
-          .setStageName(stage.name)
-          .setStageKind(stage.kind)
-          .setDirection(direction)
-      )
-      .build()
+  // Trace helpers (buildDropTrace, buildOutputTrace, packetIngressEvent, stageEvent)
+  // are shared with PSAArchitecture — see TraceHelpers.kt.
 
   // -------------------------------------------------------------------------
   // v1model extern handler
